// src/entities/management/user/types.ts
export interface UserItem {
  id: string
  full_name: string
  email: string
  phone: string
  image: string
  gender: string
  date_of_birth: string
  address: string
  department_id: string
  position_id: string
  job_title_id: string
  avatar_url: string
  hire_date: string
  refresh_token: string
  refresh_token_expiry_time: string
  is_active: string
}